refactor: 코드 구조 분리
